Add files via upload
MEDILAB pathology software belong to the class of application software intended for storage and management of information obtained in the course of the work of the laboratory. The systems are used to control and manage the information related to consultant hospitals, doctors, patient, test results, reports These services are to be provided in an efficient, cost effective manner, with the goa l of reducing the time and resources currently required for such tasks. A significant part of the operation of any Pathology Lab involves the acquisition, management and timely retrieval of great volumes of information. This information typically involves; information of consultant hospitals and doctors, patient personal information, tests information and reports for respective tests.